+++ BZip2/bzip2/BZip2OutputStream.cs
@@ -1023,7 +1023,7 @@ namespace ICSharpCode.SharpZipLib.BZip2
 						if (unLo > unHi) {
 							break;
 						}
-						n = ((int)block[zptr[unLo]+d + 1]) - med;
+						n = block[zptr[unLo] + d + 1] - med;
 						if (n == 0) {
 							int temp = zptr[unLo];
 							zptr[unLo] = zptr[ltLo];
@@ -1042,7 +1042,7 @@ namespace ICSharpCode.SharpZipLib.BZip2
 						if (unLo > unHi) {
 							break;
 						}
-						n = ((int)block[zptr[unHi]+d + 1]) - med;
+						n = block[zptr[unHi] + d + 1] - med;
 						if (n == 0) {
 							int temp = zptr[unHi];
 							zptr[unHi] = zptr[gtHi];
@@ -1126,7 +1126,7 @@ namespace ICSharpCode.SharpZipLib.BZip2
 				quadrant[i] = 0;
 			}
 			
-			block[0] = (byte)(block[last + 1]);
+			block[0] = block[last + 1];
 			
 			if (last < 4000) {
 				/*--
@@ -1303,7 +1303,7 @@ namespace ICSharpCode.SharpZipLib.BZip2
 			
 			for (i = 0; i <= last; i++) {
 				if (rNToGo == 0) {
-					rNToGo = (int)BZip2Constants.RandomNumbers[rTPos];
+					rNToGo = BZip2Constants.RandomNumbers[rTPos];
 					rTPos++;
 					if (rTPos == 512) {
 						rTPos = 0;
@@ -1549,12 +1549,12 @@ namespace ICSharpCode.SharpZipLib.BZip2
 						while (true) {
 							switch (zPend % 2) {
 								case 0:
-									szptr[wr] = (short)BZip2Constants.RunA;
+									szptr[wr] = BZip2Constants.RunA;
 									wr++;
 									mtfFreq[BZip2Constants.RunA]++;
 									break;
 								case 1:
-									szptr[wr] = (short)BZip2Constants.RunB;
+									szptr[wr] = BZip2Constants.RunB;
 									wr++;
 									mtfFreq[BZip2Constants.RunB]++;
 									break;
@@ -1577,12 +1577,12 @@ namespace ICSharpCode.SharpZipLib.BZip2
 				while (true) {
 					switch (zPend % 2) {
 						case 0:
-							szptr[wr] = (short)BZip2Constants.RunA;
+							szptr[wr] = BZip2Constants.RunA;
 							wr++;
 							mtfFreq[BZip2Constants.RunA]++;
 							break;
 						case 1:
-							szptr[wr] = (short)BZip2Constants.RunB;
+							szptr[wr] = BZip2Constants.RunB;
 							wr++;
 							mtfFreq[BZip2Constants.RunB]++;
 							break;
@@ -1709,8 +1709,8 @@ namespace ICSharpCode.SharpZipLib.BZip2
 					nNodes++;
 					parent[n1] = parent[n2] = nNodes;
 					
-					weight[nNodes] = (int)((weight[n1] & 0xffffff00) + (weight[n2] & 0xffffff00)) | 
-						(int)(1 + (((weight[n1] & 0x000000ff) > (weight[n2] & 0x000000ff)) ? (weight[n1] & 0x000000ff) : (weight[n2] & 0x000000ff)));
+					weight[nNodes] = (int)((weight[n1] & 0xffffff00) + (weight[n2] & 0xffffff00)) |
+                        1 + (((weight[n1] & 0x000000ff) > (weight[n2] & 0x000000ff)) ? (weight[n1] & 0x000000ff) : (weight[n2] & 0x000000ff));
 					
 					parent[nNodes] = -1;
 					nHeap++;
